app3/counter-service/requirements.txt
fastapi
uvicorn
pydantic
asyncpg
hazelcast-python-client==5.4.0
requests